fix(coordinator): keep complete analytics snapshots and isolate perf helpers
