+++ coordinator/store/memory.go
@@ -1231,7 +1231,7 @@ func (s *MemoryStore) addKeySpendLocked(keyID string, amount int64, at time.Time
 }
 
 // UsageLocationBuckets returns approximate request-origin aggregates (in-memory).
-func (s *MemoryStore) UsageLocationBuckets(since time.Time) []UsageLocationBucket {
+func (s *MemoryStore) UsageLocationBuckets(since time.Time) ([]UsageLocationBucket, error) {
 	s.mu.RLock()
 	defer s.mu.RUnlock()
 
@@ -1307,13 +1307,13 @@ func (s *MemoryStore) UsageLocationBuckets(since time.Time) []UsageLocationBucke
 			Providers:        len(b.providers),
 		})
 	}
-	return out
+	return out, nil
 }
 
 // UsageFlowBuckets aggregates directional consumer→provider flows in memory.
 // providerLocs supplies live provider locations from the registry; the store's
 // own providerRecords are used as a fallback for disconnected providers.
-func (s *MemoryStore) UsageFlowBuckets(since time.Time, providerLocs map[string]*ProviderLocation) []UsageFlowBucket {
+func (s *MemoryStore) UsageFlowBuckets(since time.Time, providerLocs map[string]*ProviderLocation) ([]UsageFlowBucket, error) {
 	s.mu.RLock()
 	defer s.mu.RUnlock()
 
@@ -1403,7 +1403,7 @@ func (s *MemoryStore) UsageFlowBuckets(since time.Time, providerLocs map[string]
 		}
 		out = append(out, b)
 	}
-	return out
+	return out, nil
 }
 
 // KeyCount returns the number of active API keys.
@@ -2893,28 +2893,6 @@ func (s *MemoryStore) GetProviderEarningsSummary(providerKey string) (ProviderEa
 	return summary, nil
 }
 
-// AccountEarningsWindows aggregates the account's last-24h and last-7d rows.
-func (s *MemoryStore) AccountEarningsWindows(accountID string, now time.Time) (AccountEarningsWindows, error) {
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-
-	cutoff24h := now.Add(-24 * time.Hour)
-	cutoff7d := now.Add(-7 * 24 * time.Hour)
-	var w AccountEarningsWindows
-	for _, e := range s.providerEarnings {
-		if e.AccountID != accountID || e.CreatedAt.Before(cutoff7d) {
-			continue
-		}
-		w.Last7dJobs++
-		w.Last7dMicroUSD += e.AmountMicroUSD
-		if !e.CreatedAt.Before(cutoff24h) {
-			w.Last24hJobs++
-			w.Last24hMicroUSD += e.AmountMicroUSD
-		}
-	}
-	return w, nil
-}
-
 // GetAccountEarningsSummary returns lifetime aggregates for an account.
 func (s *MemoryStore) GetAccountEarningsSummary(accountID string) (ProviderEarningsSummary, error) {
 	s.mu.RLock()
@@ -3356,22 +3334,6 @@ func (s *MemoryStore) GetReputation(_ context.Context, providerID string) (*Repu
 	}
 	cp := *rep
 	return &cp, nil
-}
-
-func (s *MemoryStore) GetReputations(_ context.Context, providerIDs []string) (map[string]*ReputationRecord, error) {
-	out := make(map[string]*ReputationRecord, len(providerIDs))
-	if len(providerIDs) == 0 {
-		return out, nil
-	}
-	s.mu.RLock()
-	defer s.mu.RUnlock()
-	for _, id := range providerIDs {
-		if rep, ok := s.reputationRecords[id]; ok {
-			cp := *rep
-			out[id] = &cp
-		}
-	}
-	return out, nil
 }
 
 // --- APNs code-identity attestation reuse cache (W5 Fix 2) ---
